Finish code to generate infeasible poses
